@@ -1,38 +1,41 @@
 package c.tlgbltcn.bluetoothhelper
 
+import android.annotation.SuppressLint
 import android.bluetooth.BluetoothDevice
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import c.tlgbltcn.bluetoothhelper.databinding.ActivityMainBinding
 import c.tlgbltcn.library.BluetoothHelper
 import c.tlgbltcn.library.BluetoothHelperListener
-import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity(), BluetoothHelperListener {
 
     private lateinit var bluetoothHelper: BluetoothHelper
-
+    private lateinit var binding: ActivityMainBinding
     private lateinit var viewAdapter: RecyclerView.Adapter<*>
 
     private var itemList = ArrayList<BluetoothDeviceModel>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        val view = binding.root
+        setContentView(view)
 
         bluetoothHelper = BluetoothHelper(this@MainActivity, this@MainActivity)
                 .setPermissionRequired(true)
                 .create()
 
-        if (bluetoothHelper.isBluetoothEnabled()) enable_disable.text = "Bluetooth State Off"
-        else enable_disable.text = "Bluetooth State On"
+        if (bluetoothHelper.isBluetoothEnabled()) binding.enableDisable.text = "Bluetooth State Off"
+        else binding.enableDisable.text = "Bluetooth State On"
 
-        if (bluetoothHelper.isBluetoothScanning()) start_stop.text = "Stop discovery"
-        else start_stop.text = "Start discovery"
+        if (bluetoothHelper.isBluetoothScanning()) binding.startStop.text = "Stop discovery"
+        else binding.startStop.text = "Start discovery"
 
 
-        enable_disable.setOnClickListener {
+        binding.enableDisable.setOnClickListener {
             if (bluetoothHelper.isBluetoothEnabled()) {
 
                 bluetoothHelper.disableBluetooth()
@@ -42,7 +45,7 @@ class MainActivity : AppCompatActivity(), BluetoothHelperListener {
             }
         }
 
-        start_stop.setOnClickListener {
+        binding.startStop.setOnClickListener {
             if (bluetoothHelper.isBluetoothScanning()) {
                 bluetoothHelper.stopDiscovery()
 
@@ -52,33 +55,33 @@ class MainActivity : AppCompatActivity(), BluetoothHelperListener {
         }
 
         viewAdapter = BluetoothListAdapter(itemList)
-        recycler_view.apply {
+        binding.recyclerView.apply {
             layoutManager = LinearLayoutManager(this@MainActivity)
             adapter = viewAdapter
         }
     }
 
     override fun onStartDiscovery() {
-        start_stop.text = "Stop discovery"
-
+        binding.startStop.text = "Stop discovery"
     }
 
     override fun onFinishDiscovery() {
-        start_stop.text = "Start discovery"
+        binding.startStop.text = "Start discovery"
         itemList.clear()
     }
 
     override fun onEnabledBluetooth() {
-        enable_disable.text = "Bluetooth State Off"
+        binding.enableDisable.text = "Bluetooth State Off"
     }
 
     override fun onDisabledBluetooh() {
-        enable_disable.text = "Bluetooth State On"
+        binding.enableDisable.text = "Bluetooth State On"
 
     }
 
-    override fun getBluetoothDeviceList(device: BluetoothDevice) {
-        itemList.add(BluetoothDeviceModel(device.name, device.address))
+    @SuppressLint("NotifyDataSetChanged")
+    override fun getBluetoothDeviceList(device: BluetoothDevice?) {
+        itemList.add(BluetoothDeviceModel(device?.name, device?.address))
         viewAdapter.notifyDataSetChanged()
     }
 
